[#76]refact:위치값기준 주소뽑아오는부분 파일분리
