+++ src/components/stamp/MyLocation.tsx
@@ -1,12 +1,7 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import StampActive from './StampActive';
-import { AddressPropsType } from '@/types/stamp/AddressProps.types';
-import { showErrorMsg } from '@/components/stamp/LocationErrorMsg';
-
-import browserClient from '@/utils/supabase/client';
-import { useQuery } from '@tanstack/react-query';
+import StampActive from '@/components/stamp/StampActive';
 import useUser from '@/hooks/useUser';
 import Link from 'next/link';
 import Icon from '@/components/common/Icons/Icon';
@@ -15,53 +10,22 @@ import useModal from '@/hooks/useModal';
 import AliasCheckModal from '@/components/common/Modal/AliasCheckModal';
 import { useGetStampLocationQuery } from '@/queries/query/useStampQuery';
 import { usePatchAliasMutation } from '@/queries/mutation/useAliasMutation';
-import { getStampLocation } from '@/components/stamp/getStampLocation';
-
-interface LocationType {
-  lat: number;
-  lng: number;
-}
+import GetUserAddress from '@/components/stamp/GetUserAddress';
 
 const MyLocation = () => {
-  const userId = useUser();
-  const [error, setError] = useState<string | null>(null); //에러상태
-  const [address, setAddress] = useState<AddressPropsType>(); //현재주소
+  const userId = useUser(); //유저아이디
   const [visit, setVisit] = useState<boolean>(false); //방문상태
-  const [location, setLocation] = useState<LocationType>({ lat: 0, lng: 0 });
-  const [parentFocused, setParentFocused] = useState(false);
+  const [parentFocused, setParentFocused] = useState(false); //포커스상태
   const [aliasLocation, setAliasLocation] = useState<string | null>(null); //장소별칭
-  const { openModal, Modal, isOpen } = useModal();
-  const patchAliasMutation = usePatchAliasMutation();
-  // const [saveAlias,setSaveAlias] = useState(stampList?.[0]?.aliasLocation)
+  const { openModal, Modal, isOpen } = useModal(); //모달훅
+  const postAliasMutation = usePatchAliasMutation();
 
-  // const fetchLocationStamp = async (address: string) => {
-  //   const { data: userLocationStamp, error } = await browserClient
-  //     .from('stamp')
-  //     .select('*')
-  //     .eq('user_id', userId)
-  //     .eq('address', address);
-
-  //   if (error) {
-  //     console.error('위치 기반 스탬프 리스트 가져오기 오류 :', error.message);
-  //     throw new Error('위치 기반 스탬프 리스트 데이터를 가져오는 중 오류가 발생했습니다.' + error.message);
-  //   }
-  //   return userLocationStamp;
-  // };
-
+  const { address, location, error } = GetUserAddress();
   const {
     data: stampList,
     isLoading,
     isError: stampListError
   } = useGetStampLocationQuery(address?.address_name, userId);
-  // useQuery({
-  //   queryKey: ['nowStamp', address?.address_name], //고유키값
-  //   queryFn: async () => {
-  //     if (address && address.address_name) {
-  //       return await getStampLocation(address.address_name!);
-  //     } else return null;
-  //   }, // 주소를 인자로 넘김
-  //   enabled: !!userId
-  // });
 
   //저장된스탬프의 방문여부 저장
   useEffect(() => {
@@ -70,69 +34,24 @@ const MyLocation = () => {
     }
   }, [stampList]);
 
+  //스탬프 별명 추가 이벤트
   const onClickAliasAdd = (alias: string) => {
     if (userId) {
-      patchAliasMutation.mutate({ alias, userId, address: address?.address_name ?? '' });
+      postAliasMutation.mutate({ alias, userId, address: address?.address_name ?? '' });
     } else {
       console.error('유저아이디가 없습니다.');
       return;
     }
   };
 
+  // 스탬프 별명있을때 실행
   useEffect(() => {
     if (aliasLocation) {
       onClickAliasAdd(aliasLocation);
     }
   }, [aliasLocation]);
 
-  // 카카오맵 주소값 가져오기
-  const getAddress = async (lat: number, lng: number) => {
-    try {
-      const res = await fetch(`https://dapi.kakao.com/v2/local/geo/coord2address.json?x=${lng}&y=${lat}`, {
-        headers: {
-          Authorization: `KakaoAK ${process.env.NEXT_PUBLIC_KAKAO_REST_API_KEY}`
-        }
-      });
-      const data = await res.json();
-      if (data) {
-        const addressData = data.documents[0].address;
-        setAddress({
-          address_name: addressData.address_name,
-          region_1depth_name: addressData.region_1depth_name,
-          region_2depth_name: addressData.region_2depth_name,
-          region_3depth_name: addressData.region_3depth_name
-        });
-      }
-    } catch (error) {
-      console.log('주소를 가져오는 중 오류가 발생했습니다.', error);
-    }
-  };
-
-  // Geolocation API 로 유저의 위도,경도값 추출
-  useEffect(() => {
-    if ('geolocation' in navigator) {
-      //현 브라우저가 Geolocation API를 지원하는지 확인
-      navigator.geolocation.getCurrentPosition(
-        //사용자의 현재 위치를 요청
-        async (position) => {
-          const { latitude, longitude } = position.coords;
-          setLocation({ lat: latitude, lng: longitude });
-          await getAddress(latitude, longitude);
-        },
-        (err) => {
-          showErrorMsg(err.message, setError);
-        },
-        {
-          enableHighAccuracy: true, // 정확도 우선 모드
-          timeout: 60000, // 1분 이내에 응답 없으면 에러 발생
-          maximumAge: 0 // 항상 최신 위치 정보 수집
-        }
-      );
-    } else {
-      console.log('오류가 발생했습니다.');
-    }
-  }, []);
-
+  //TODO: 추후 지우기
   useEffect(() => {
     if (isLoading) {
       console.log('1');
@@ -151,8 +70,9 @@ const MyLocation = () => {
     );
   }
 
+  //스탬프찍힌 위치의 별명 저장
   const saveAlias = stampList?.[0]?.aliasLocation;
-  // console.log('stampList', stampList?.[0]?.aliasLocation);
+
   return (
     <div className="flex flex-col px-[24px] pt-[36px]" style={{ height: 'calc(100vh - 64px)' }}>
       {address ? (
@@ -184,7 +104,6 @@ const MyLocation = () => {
                 color={`${parentFocused ? '#00688A' : '#9C9C9C'}`}
                 bgColor="transparent"
               />
-              {/* TODO: 등록된거라면 별칭 보여주기 */}
               <input
                 type="text"
                 placeholder={`${stampList?.[0]?.aliasLocation ? stampList?.[0]?.aliasLocation : '간단한 장소나 이름을 적어주세요.'}`}
